+++ utils/src/main/scala/de.welt.contentapi.utils/Strings.scala
@@ -1,4 +1,4 @@
-package de.welt.contentapi.core.client.utilities
+package de.welt.contentapi.utils
 
 import com.google.common.base.CharMatcher
 
@@ -9,19 +9,19 @@ trait Strings {
   /**
     * remove all whitespaces from the string (such as {{{"\u00A0", " ", "\t", "\n"}}})
     */
-  val stripWhiteSpaces = (s: String) ⇒ charMatcher.removeFrom(s)
+  val stripWhiteSpaces: (String) ⇒ String = (s: String) ⇒ charMatcher.removeFrom(s)
 
   /**
     * `true` if `s` contains only whitespaces
     * <br/> `false` otherwise
     */
-  val containsOnlyWhitespaces = (s: String) ⇒ charMatcher.matchesAllOf(s)
+  val containsOnlyWhitespaces: (String) ⇒ Boolean = (s: String) ⇒ charMatcher.matchesAllOf(s)
 
   /**
     * `true` if `s` contains any non-whitespaces
     * <br/> `false` otherwise
     */
-  val containsTextContent = (s: String) ⇒ !containsOnlyWhitespaces(s)
+  val containsTextContent: (String) ⇒ Boolean = (s: String) ⇒ !containsOnlyWhitespaces(s)
 
 }